server/requirements.lock.txt
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.1.3
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dj-rest-auth==4.0.1
Django==4.2.2
django-allauth==0.54.0
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
executing==1.2.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jwcrypto==1.5.0
Markdown==3.4.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.7.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
swagger-spec-validator==3.0.3
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.6
wrapt==1.15.0
